// LinkERP/ClientApp/src/app/models/inv/lbs-inv-inventory-vendor.ts
import { BaseEntity } from "../base-entity";

export class LBSINVInventoryVendor extends BaseEntity {
    CompanyID: any;
    InventoryID: any;
    VendorID: any;
    PurchaseUOM: any;
    VendorBarCode: any;
    Default: any;
    LastPurchasedDate: any;
    DateCreated: any;
    LastPurchasedUnitPrice: any;
    VendorSKU: any;
    VendorAccountName:any;
    VendorWareHouse:any;
}